First simulation performed - Added references and notes
Created a gmm using the parameters' statistical description and generated a random walk to try the simulator. Ocassional problems on negative weights and assymetrical Sigma to be fixed. Also added notes and referenced projects/papers as well as a test to see if Martina Fornaciari's work corresponds to my data